requirements.txt
beautifulsoup4==4.12.2
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6
requests==2.31.0
soupsieve==2.5
urllib3==2.1.0
uvicorn==0.24.0
python-dotenv==1.0.0
fastapi==0.104.1
backoff==2.2.1
openai==1.3.7
langchain==0.0.345
langdetect==1.0.9